Fix/remitos kg decimal y edicion

+++ src/main/java/com/tulumcore/api/controllers/ProductoDTO.java
@@ -6,7 +6,7 @@ public class ProductoDTO {
     private Double precio;
     private Double precioCosto;
     private Double margenPorcentaje;
-    private Integer cantidadStock;
+    private Double cantidadStock;
     private Integer stockMinimo;
     private String medidas;
     private String codigoBarras;
@@ -23,8 +23,8 @@ public class ProductoDTO {
     public void setPrecioCosto(Double precioCosto) { this.precioCosto = precioCosto; }
     public Double getMargenPorcentaje() { return margenPorcentaje; }
     public void setMargenPorcentaje(Double margenPorcentaje) { this.margenPorcentaje = margenPorcentaje; }
-    public Integer getCantidadStock() { return cantidadStock; }
-    public void setCantidadStock(Integer cantidadStock) { this.cantidadStock = cantidadStock; }
+    public Double getCantidadStock() { return cantidadStock; }
+    public void setCantidadStock(Double cantidadStock) { this.cantidadStock = cantidadStock; }
     public Integer getStockMinimo() { return stockMinimo; }
     public void setStockMinimo(Integer stockMinimo) { this.stockMinimo = stockMinimo; }
     public String getMedidas() { return medidas; }
